+++ client/src/commons/types/calendar.ts
@@ -1,0 +1,1 @@
+export type SelectDayModel = { day: number };
