+++ client/src/commons/types/date.ts
@@ -1,3 +1,1 @@
 export type YearMonthModel = { year: number; month: number };
-
-export type SelectDayModel = { day: number };
