+++ client/src/libs/CalendarUtils.ts
@@ -26,6 +26,7 @@ const getDayMatrix = (year: number, month: number): string[][] => {
       result.push(dataArr.slice(i * weekSlice, i * weekSlice + weekSlice));
       i += 1;
     }
+    if (result.length > 5) result.pop();
     return result;
   };
   return resultMatrix();
