+++ client/src/modules/calendarDaySelector/action.ts
@@ -1,0 +1,13 @@
+import { SelectDayModel } from '@/commons/types/calendar';
+
+type ChangeDayActionType = {
+  type: string;
+  payload: SelectDayModel;
+};
+
+export const CHANGE_DAY = 'day/CHANGE_DAY';
+
+export const changeDay = (date: SelectDayModel): ChangeDayActionType => ({
+  type: CHANGE_DAY,
+  payload: date,
+});
