+++ client/src/modules/calendarDaySelector/type.ts
@@ -1,0 +1,7 @@
+import { SelectDayModel } from '@/commons/types/calendar';
+import { ActionType } from 'typesafe-actions';
+import * as actions from './action';
+
+export type SelectDayAction = ActionType<typeof actions>;
+
+export type SelectDayState = SelectDayModel;
