+++ client/src/modules/datePicker/index.ts
@@ -1,3 +1,3 @@
-export { datePicker, dayPicker } from './reducer';
+export { default } from './reducer';
 export * from './actions';
 export * from './types';
