+++ client/src/modules/datePicker/types.ts
@@ -1,11 +1,7 @@
-import { YearMonthModel, SelectDayModel } from '@/commons/types/date';
+import { YearMonthModel } from '@/commons/types/date';
 import { ActionType } from 'typesafe-actions';
 import * as actions from './actions';
 
 export type DatePickerAction = ActionType<typeof actions>;
 
 export type DatePickerState = YearMonthModel;
-
-export type DayPickerAction = ActionType<typeof actions>;
-
-export type DayPickerState = SelectDayModel;
